requirements.txt
requests==2.1.0
beautifulsoup4==4.3.2
python-dateutil==2.1
lxml==3.2.4
